+++ contracts/src/v0.8/keystone/test/CapabilityRegistry_UpdateNodesTest.t.sol
@@ -7,7 +7,7 @@ import {CapabilityRegistry} from "../CapabilityRegistry.sol";
 contract CapabilityRegistry_UpdateNodesTest is BaseTest {
   event NodeUpdated(bytes32 p2pId, uint256 nodeOperatorId, address signer);
 
-  uint256 private constant TEST_NODE_OPERATOR_ONE_ID = 0;
+  uint32 private constant TEST_NODE_OPERATOR_ONE_ID = 0;
   uint256 private constant TEST_NODE_OPERATOR_TWO_ID = 1;
   bytes32 private constant INVALID_P2P_ID = bytes32("fake-p2p");
